updating the read unread mentions from room intent handler
